refactor aws ses and add initial mailgun manager
